twitter_app/requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
djangorestframework==3.7.7
future==0.16.0
idna==2.6
oauthlib==2.0.6
python-dateutil==2.6.1
python-twitter==3.4
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tweepy==3.5.0
twitter==1.18.0
urllib3==1.22
